Silence logger reduction warning related to integer metrics

+++ point_frustums/metrics/detection/featuremap_level_assignments.py
@@ -6,7 +6,7 @@ class FeaturemapLevelAssignments(Metric):
     # pylint: disable=no-member
     def __init__(self, num_levels):
         super().__init__()
-        self.add_state("level_assignment_count", default=torch.zeros(num_levels).int(), dist_reduce_fx="sum")
+        self.add_state("level_assignment_count", default=torch.zeros(num_levels).float(), dist_reduce_fx="sum")
 
     def update(self, count) -> None:  # pylint: disable=arguments-differ
         self.level_assignment_count += count
@@ -19,7 +19,7 @@ class PreassignmentOutflows(Metric):
     # pylint: disable=no-member
     def __init__(self):
         super().__init__()
-        self.add_state("outflow", default=torch.zeros(1).int(), dist_reduce_fx="sum")
+        self.add_state("outflow", default=torch.zeros(1).float(), dist_reduce_fx="sum")
 
     def update(self, count) -> None:  # pylint: disable=arguments-differ
         self.outflow += count
@@ -32,7 +32,7 @@ class MissedTargets(Metric):
     # pylint: disable=no-member
     def __init__(self):
         super().__init__()
-        self.add_state("missed_targets", default=torch.zeros(1).int(), dist_reduce_fx="sum")
+        self.add_state("missed_targets", default=torch.zeros(1).float(), dist_reduce_fx="sum")
 
     def update(self, count: int | torch.Tensor) -> None:  # pylint: disable=arguments-differ
         self.missed_targets += count
